+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @token-icons/core
 
+## 2.7.1
+
+### Patch Changes
+
+- new tokens: SENSE, AMZ, USN
+
+## 2.7.0
+
+### Minor Changes
+
+- add network icons: Zora, XAI
+
 ## 2.6.1
 
 ### Patch Changes